trailpack-react-ui/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.18
pandas==2.2.3
pyarrow==18.1.0
openpyxl==3.1.5
httpx==0.27.2
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
